+++ releases/changes.md
@@ -1,2 +1,4 @@
 * Improved STDIN messages, that can be disabled by `$SEQFU_QUIET=1`
+* Added `--format irida` in `seqfu metadata` (for [IRIDA uploader](https://github.com/phac-nml/irida-uploader))
+* Added `--gc` in `seqfu qual`: will print an additional column
 * Minor improvements on `seqfu cat`